+++ 0x1E-search_algorithms/search_algos.h
@@ -8,6 +8,6 @@ void print_array(int *array, int l, int r);
 int do_binary_search(int *array, size_t size, int value);
 int jump_search(int *array, size_t size, int value);
 int continue_search(int *array, size_t size
-		, int value, size_t currentIdx, size_t jump);
+		, int value, int currentIdx, int jump);
 
 #endif
